requirements.txt
pytesseract 
pillow 
PyPDF2 
reportlab 
openai
pdf2image
python-dotenv